+++ src/HetrickCV.cpp
@@ -38,6 +38,7 @@ void init(rack::Plugin *p) {
 	p->addModel(modelLogicCombine);
 	p->addModel(modelMidSide);
 	p->addModel(modelMinMax);
+	p->addModel(modelPhasorGen);
 	p->addModel(modelProbability);
 	p->addModel(modelRandomGates);
 	p->addModel(modelRotator);